requirements/requirements.txt
alabaster==0.7.16
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
attrs==23.2.0
Babel==2.15.0
binaryornot==0.4.4
boto3==1.34.117
botocore==1.34.117
certifi==2024.6.2
cffi==1.16.0
cfgv==3.4.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cookiecutter==2.6.0
coverage==7.5.3
crispy-bootstrap5==2024.2
cssbeautifier==1.15.1
cycler==0.12.1
decorator==5.1.1
distlib==0.3.8
Django==4.2.13
django-allauth==0.63.2
django-anymail==10.3
django-appconf==1.0.6
django-compressor==4.4
django-cors-headers==4.3.1
django-coverage-plugin==3.1.0
django-crispy-forms==2.1
django-debug-toolbar==4.4.2
django-environ==0.11.2
django-extensions==3.2.3
django-model-utils==4.5.1
django-redis==5.4.0
django-storages==1.14.3
django-stubs==5.0.2
django-stubs-ext==5.0.2
djangorestframework==3.15.1
djangorestframework-stubs==3.15.0
djlint==1.34.1
docutils==0.21.2
drf-spectacular==0.27.2
EditorConfig==0.12.4
executing==2.0.1
factory-boy==3.3.0
Faker==25.3.0
filelock==3.14.0
fonttools==4.53.0
fsspec==2024.5.0
gunicorn==22.0.0
h11==0.14.0
hiredis==2.3.2
html-tag-names==0.1.2
html-void-elements==0.1.0
identify==2.5.36
idna==3.7
imagesize==1.4.1
inflection==0.5.1
iniconfig==2.0.0
ipdb==0.13.13
ipython==8.25.0
jedi==0.19.1
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
jsbeautifier==1.15.1
json5==0.9.25
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
mypy==1.10.0
mypy-extensions==1.0.0
networkx==3.3
nodeenv==1.9.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
packaging==24.0
pandas==2.2.2
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.2
pluggy==1.5.0
pre-commit==3.7.1
prompt_toolkit==3.0.45
psycopg==3.1.19
psycopg-c==3.1.19
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.22
Pygments==2.18.0
pyparsing==3.1.2
pypng==0.20220715.0
pytest==8.2.1
pytest-django==4.8.0
pytest-sugar==1.0.0
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.1
qrcode==7.4.2
rcssmin==1.1.1
redis==5.0.4
referencing==0.35.1
regex==2023.12.25
requests==2.32.3
rich==13.7.1
rjsmin==1.2.1
rpds-py==0.18.1
ruff==0.4.7
s3transfer==0.10.1
scikit-learn==1.5.0
scipy==1.13.1
six==1.16.0
sniffio==1.3.1
snowballstemmer==2.2.0
Sphinx==7.3.7
sphinx-autobuild==2024.4.16
sphinxcontrib-applehelp==1.0.8
sphinxcontrib-devhelp==1.0.6
sphinxcontrib-htmlhelp==2.0.5
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.7
sphinxcontrib-serializinghtml==1.1.10
sqlparse==0.5.0
stack-data==0.6.3
starlette==0.37.2
sympy==1.12.1
termcolor==2.4.0
text-unidecode==1.3
threadpoolctl==3.5.0
torch==2.3.0
torchvision==0.18.0
tqdm==4.66.4
traitlets==5.14.3
triton==2.3.0
types-python-dateutil==2.9.0.20240316
types-PyYAML==6.0.12.20240311
types-requests==2.32.0.20240602
typing_extensions==4.12.1
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.30.1
virtualenv==20.26.2
watchdog==4.0.1
watchfiles==0.22.0
wcwidth==0.2.13
websockets==12.0
Werkzeug==3.0.3
whitenoise==6.6.0
